+++ fake-snippets-api/lib/db/autoload-snippets.json
@@ -7,6 +7,10 @@
     {
       "owner": "ShiboSoftwareDev",
       "name": "Wifi-Camera-Module"
+    },
+    {
+      "owner": "imrishabh18",
+      "name": "Arduino-nano"
     }
   ]
 }
